Refactored enums by moving to a separate package
