+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/enums/BookSortBy.kt
@@ -1,0 +1,6 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.enums
+
+enum class BookSortBy {
+    BOOK_ORDER,
+    ALPHABETICAL
+}
