+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/enums/ProjectType.kt
@@ -1,0 +1,6 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.enums
+
+enum class ProjectType(val value: String) {
+    DRAFTING("draftingProject"),
+    FINAL_RECORDING("finalRecordingProject")
+}
